dev: Forgot in last commit, added logic to decrease comment count by one on successful comment delete.

+++ src/app/features/posts/components/comment/comment.component.ts
@@ -33,6 +33,12 @@ export class CommentComponent implements OnInit {
 
     deleteComment() {
 
+        const accepted = confirm('Are you sure you want to delete this comment?');
+
+        if (!accepted) {
+            return;
+        }
+
         this.commentService.deleteComment(this.data().commentId).pipe(takeUntil(this.destroy$)).subscribe({
             next: () => {
                 this.commentDeleted.emit(this.data().commentId);
